+++ test/view/toolbar.test.ts
@@ -241,7 +241,6 @@ describe('toolbar controls', () => {
 			'New Idea',
 			'New Deliverable',
 			'New Milestone',
-			'New Iteration',
 		]);
 
 		picker?.item('New PBI')?.click();